move event chunk and gatway contract class structs to version folder (#846)
* move contract class tests to its one file

* move event chunk and gatway contract class structs to version folder
